BREAKING: drop --light option
