+++ Makefile
@@ -46,18 +46,13 @@ clean:
 $(GOBIN)/bbolt:
 	go get -v go.etcd.io/bbolt/...
 
-export DB_TYPE ?= trivy
-
-ifeq ($(DB_TYPE),trivy-light)
-	DB_ARG := --light
-endif
-
 trivy-db:
 	make build
 
 .PHONY: db-all
 db-all:
-	make build db-fetch-langs db-fetch-vuln-list-main
+	#make build db-fetch-langs db-fetch-vuln-list-main
+	make build
 	make db-build
 	make db-compact
 	make db-compress
@@ -73,19 +68,18 @@ db-fetch-langs:
 
 .PHONY: db-build
 db-build: trivy-db
-	./trivy-db build $(DB_ARG) --cache-dir cache --update-interval 6h
+	./trivy-db build --cache-dir cache --update-interval 6h
 
 .PHONY: db-compact
 db-compact: $(GOBIN)/bbolt cache/db/trivy.db
-	mkdir -p assets/$(DB_TYPE)
-	$(GOBIN)/bbolt compact -o ./assets/$(DB_TYPE)/$(DB_TYPE).db cache/db/trivy.db
-	cp cache/db/metadata.json ./assets/$(DB_TYPE)/metadata.json
+	mkdir assets/
+	$(GOBIN)/bbolt compact -o ./assets/trivy.db cache/db/trivy.db
+	cp cache/db/metadata.json ./assets/metadata.json
 	rm cache/db/trivy.db
 
 .PHONY: db-compress
-db-compress: assets/$(DB_TYPE)/$(DB_TYPE).db assets/$(DB_TYPE)/metadata.json
-	tar cvzf assets/$(DB_TYPE)-offline.db.tgz -C assets/$(DB_TYPE) $(DB_TYPE).db metadata.json
-	gzip --best -c assets/$(DB_TYPE)/$(DB_TYPE).db > assets/$(DB_TYPE).db.gz
+db-compress: assets/trivy.db assets/metadata.json
+	tar cvzf assets/db.tar.gz -C assets/ trivy.db metadata.json
 
 .PHONY: db-clean
 db-clean:
