+++ pkg/build.go
@@ -11,19 +11,13 @@ import (
 func build(c *cli.Context) error {
 	cacheDir := c.String("cache-dir")
 	if err := db.Init(cacheDir); err != nil {
-		return err
+		return xerrors.Errorf("db initialize error: %w", err)
 	}
 
 	targets := c.StringSlice("only-update")
-	light := c.Bool("light")
 	updateInterval := c.Duration("update-interval")
 
-	dbType := db.TypeFull
-	if light {
-		dbType = db.TypeLight
-	}
-
-	vdb := vulndb.New(dbType, cacheDir, updateInterval)
+	vdb := vulndb.New(cacheDir, updateInterval)
 	if err := vdb.Build(targets); err != nil {
 		return xerrors.Errorf("build error: %w", err)
 	}
