+++ pkg/db/db.go
@@ -18,13 +18,8 @@ import (
 
 type CustomPut func(dbc Operation, tx *bolt.Tx, adv interface{}) error
 
-type Type int
-
 const (
 	SchemaVersion = 1
-
-	TypeFull Type = iota
-	TypeLight
 )
 
 var (
@@ -45,11 +40,8 @@ type Operation interface {
 	ForEachAdvisory(source string, pkgName string) (value map[string][]byte, err error)
 	GetAdvisories(source string, pkgName string) (advisories []types.Advisory, err error)
 
-	PutSeverity(tx *bolt.Tx, vulnerabilityID string, severity types.Severity) (err error)
-	GetSeverity(vulnerabilityID string) (severity types.Severity, err error)
-	ForEachSeverity(fn func(tx *bolt.Tx, cveID string, severity types.Severity) error) (err error)
-
-	DeleteSeverityBucket() (err error)
+	PutVulnerabilityID(tx *bolt.Tx, vulnerabilityID string) (err error)
+	ForEachVulnerabilityID(fn func(tx *bolt.Tx, cveID string) error) (err error)
 
 	PutVulnerability(tx *bolt.Tx, vulnerabilityID string, vulnerability types.Vulnerability) (err error)
 	GetVulnerability(vulnerabilityID string) (vulnerability types.Vulnerability, err error)
@@ -61,8 +53,7 @@ type Operation interface {
 }
 
 type Metadata struct {
-	Version      int  `json:",omitempty"`
-	Type         Type `json:",omitempty"`
+	Version      int `json:",omitempty"`
 	NextUpdate   time.Time
 	UpdatedAt    time.Time
 	DownloadedAt time.Time
