+++ pkg/vulndb/db.go
@@ -5,6 +5,7 @@ import (
 	"path/filepath"
 	"time"
 
+	bolt "go.etcd.io/bbolt"
 	"golang.org/x/xerrors"
 	"k8s.io/utils/clock"
 
@@ -17,7 +18,7 @@ type VulnDB interface {
 	Build(targets []string) error
 }
 
-type Core struct {
+type TrivyDB struct {
 	dbc            db.Config
 	vulnClient     vulnerability.Vulnerability
 	vulnSrcs       map[string]vulnsrc.VulnSrc
@@ -26,21 +27,21 @@ type Core struct {
 	clock          clock.Clock
 }
 
-type Option func(*Core)
+type Option func(*TrivyDB)
 
 func WithClock(clock clock.Clock) Option {
-	return func(core *Core) {
+	return func(core *TrivyDB) {
 		core.clock = clock
 	}
 }
 
 func WithVulnSrcs(srcs map[string]vulnsrc.VulnSrc) Option {
-	return func(core *Core) {
+	return func(core *TrivyDB) {
 		core.vulnSrcs = srcs
 	}
 }
 
-func NewCore(cacheDir string, updateInterval time.Duration, opts ...Option) *Core {
+func New(cacheDir string, updateInterval time.Duration, opts ...Option) *TrivyDB {
 	// Initialize map
 	vulnSrcs := map[string]vulnsrc.VulnSrc{}
 	for _, v := range vulnsrc.All {
@@ -48,7 +49,7 @@ func NewCore(cacheDir string, updateInterval time.Duration, opts ...Option) *Cor
 	}
 
 	dbc := db.Config{}
-	core := &Core{
+	tdb := &TrivyDB{
 		dbc:            dbc,
 		vulnClient:     vulnerability.New(dbc),
 		vulnSrcs:       vulnSrcs,
@@ -58,39 +59,38 @@ func NewCore(cacheDir string, updateInterval time.Duration, opts ...Option) *Cor
 	}
 
 	for _, opt := range opts {
-		opt(core)
+		opt(tdb)
 	}
 
-	return core
+	return tdb
 }
 
-func (c Core) Insert(dbType db.Type, targets []string) error {
+func (t TrivyDB) Insert(targets []string) error {
 	log.Println("Updating vulnerability database...")
 	for _, target := range targets {
-		src, ok := c.vulnSrc(target)
+		src, ok := t.vulnSrc(target)
 		if !ok {
 			return xerrors.Errorf("%s is not supported", target)
 		}
 		log.Printf("Updating %s data...\n", target)
 
-		if err := src.Update(c.cacheDir); err != nil {
+		if err := src.Update(t.cacheDir); err != nil {
 			return xerrors.Errorf("%s update error: %w", target, err)
 		}
 	}
 
 	md := db.Metadata{
 		Version:    db.SchemaVersion,
-		Type:       dbType,
-		NextUpdate: c.clock.Now().UTC().Add(c.updateInterval),
-		UpdatedAt:  c.clock.Now().UTC(),
+		NextUpdate: t.clock.Now().UTC().Add(t.updateInterval),
+		UpdatedAt:  t.clock.Now().UTC(),
 	}
 
-	err := c.dbc.SetMetadata(md)
+	err := t.dbc.SetMetadata(md)
 	if err != nil {
 		return xerrors.Errorf("failed to save metadata: %w", err)
 	}
 
-	err = c.dbc.StoreMetadata(md, filepath.Join(c.cacheDir, "db"))
+	err = t.dbc.StoreMetadata(md, filepath.Join(t.cacheDir, "db"))
 	if err != nil {
 		return xerrors.Errorf("failed to store metadata: %w", err)
 	}
@@ -98,8 +98,27 @@ func (c Core) Insert(dbType db.Type, targets []string) error {
 	return nil
 }
 
-func (c Core) vulnSrc(target string) (vulnsrc.VulnSrc, bool) {
-	for _, src := range c.vulnSrcs {
+func (t TrivyDB) Build(targets []string) error {
+	// Insert all security advisories
+	if err := t.Insert(targets); err != nil {
+		return xerrors.Errorf("insert error: %w", err)
+	}
+
+	// Remove unnecessary details
+	if err := t.optimize(); err != nil {
+		return xerrors.Errorf("optimize error: %w", err)
+	}
+
+	// Remove unnecessary buckets
+	if err := t.cleanup(); err != nil {
+		return xerrors.Errorf("cleanup error: %w", err)
+	}
+
+	return nil
+}
+
+func (t TrivyDB) vulnSrc(target string) (vulnsrc.VulnSrc, bool) {
+	for _, src := range t.vulnSrcs {
 		if target == src.Name() {
 			return src, true
 		}
@@ -107,13 +126,44 @@ func (c Core) vulnSrc(target string) (vulnsrc.VulnSrc, bool) {
 	return nil, false
 }
 
-func New(dbType db.Type, cacheDir string, updateInterval time.Duration, opts ...Option) VulnDB {
-	core := NewCore(cacheDir, updateInterval, opts...)
+func (t TrivyDB) optimize() error {
+	err := t.dbc.ForEachVulnerabilityID(func(tx *bolt.Tx, cveID string) error {
+		details := t.vulnClient.GetDetails(cveID)
+		if t.vulnClient.IsRejected(details) {
+			return nil
+		}
 
-	switch dbType {
-	case db.TypeLight:
-		return lightDB{Core: core}
-	default:
-		return fullDB{Core: core}
+		if err := t.vulnClient.SaveAdvisoryDetails(tx, cveID); err != nil {
+			return xerrors.Errorf("failed to save advisories: %w", err)
+		}
+
+		vuln := t.vulnClient.Normalize(details)
+		if err := t.dbc.PutVulnerability(tx, cveID, vuln); err != nil {
+			return xerrors.Errorf("failed to put vulnerability: %w", err)
+		}
+
+		return nil
+	})
+
+	if err != nil {
+		return xerrors.Errorf("failed to iterate severity: %w", err)
 	}
+
+	return nil
+}
+
+func (t TrivyDB) cleanup() error {
+	if err := t.dbc.DeleteVulnerabilityIDBucket(); err != nil {
+		return xerrors.Errorf("failed to delete severity bucket: %w", err)
+	}
+
+	if err := t.dbc.DeleteVulnerabilityDetailBucket(); err != nil {
+		return xerrors.Errorf("failed to delete vulnerability detail bucket: %w", err)
+	}
+
+	if err := t.dbc.DeleteAdvisoryDetailBucket(); err != nil {
+		return xerrors.Errorf("failed to delete advisory detail bucket: %w", err)
+	}
+
+	return nil
 }
